requirements.txt
transformers
python-pptx
flask
